service: flatten azure controller
