+++ service/controller/azure/error.go
@@ -1,8 +1,6 @@
 package azure
 
-import (
-	"github.com/giantswarm/microerror"
-)
+import "github.com/giantswarm/microerror"
 
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
