+++ service/controller/azure/resource/chartconfig/current.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/operatorkit/controller/context/resourcecanceledcontext"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
-	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
+	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/key"
 	"github.com/giantswarm/cluster-operator/service/controller/internal/chartconfig"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
 )
