+++ service/controller/azure/resource/chartconfig/desired.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/giantswarm/microerror"
 
-	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
+	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/key"
 	"github.com/giantswarm/cluster-operator/service/controller/internal/chartconfig"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
 )
