+++ service/controller/azure/resource/chartconfig/update.go
@@ -1,4 +1,4 @@
-package configmap
+package chartconfig
 
 import (
 	"context"
@@ -8,8 +8,8 @@ import (
 	"github.com/giantswarm/operatorkit/controller"
 	"github.com/giantswarm/operatorkit/controller/context/reconciliationcanceledcontext"
 
-	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
-	"github.com/giantswarm/cluster-operator/service/controller/internal/configmap"
+	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/key"
+	"github.com/giantswarm/cluster-operator/service/controller/internal/chartconfig"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
 )
 
@@ -19,7 +19,7 @@ func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange inte
 		return microerror.Mask(err)
 	}
 
-	configMapsToUpdate, err := toConfigMaps(updateChange)
+	chartConfigsToUpdate, err := toChartConfigs(updateChange)
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -30,11 +30,13 @@ func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange inte
 		return microerror.Mask(err)
 	}
 
-	clusterConfig := configmap.ClusterConfig{
-		APIDomain: apiDomain,
-		ClusterID: key.ClusterID(clusterGuestConfig),
+	clusterConfig := chartconfig.ClusterConfig{
+		APIDomain:    apiDomain,
+		ClusterID:    key.ClusterID(clusterGuestConfig),
+		Organization: key.ClusterOrganization(clusterGuestConfig),
 	}
-	err = r.configMap.ApplyUpdateChange(ctx, clusterConfig, configMapsToUpdate)
+
+	err = r.chartConfig.ApplyUpdateChange(ctx, clusterConfig, chartConfigsToUpdate)
 	if tenant.IsAPINotAvailable(err) {
 		r.logger.LogCtx(ctx, "level", "debug", "message", "tenant cluster is not available")
 
@@ -52,17 +54,17 @@ func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange inte
 }
 
 func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*controller.Patch, error) {
-	currentConfigMaps, err := toConfigMaps(currentState)
+	currentChartConfigs, err := toChartConfigs(currentState)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	desiredConfigMaps, err := toConfigMaps(desiredState)
+	desiredChartConfigs, err := toChartConfigs(desiredState)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	patch, err := r.configMap.NewUpdatePatch(ctx, currentConfigMaps, desiredConfigMaps)
+	patch, err := r.chartConfig.NewUpdatePatch(ctx, currentChartConfigs, desiredChartConfigs)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
