+++ service/controller/azure/resource_set.go
@@ -1,4 +1,4 @@
-package v22
+package azure
 
 import (
 	"context"
@@ -22,9 +22,9 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/giantswarm/cluster-operator/pkg/label"
-	"github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
-	"github.com/giantswarm/cluster-operator/service/controller/azure/v22/resource/chartconfig"
-	"github.com/giantswarm/cluster-operator/service/controller/azure/v22/resource/configmap"
+	"github.com/giantswarm/cluster-operator/service/controller/azure/key"
+	"github.com/giantswarm/cluster-operator/service/controller/azure/resource/chartconfig"
+	"github.com/giantswarm/cluster-operator/service/controller/azure/resource/configmap"
 	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
 	chartconfigservice "github.com/giantswarm/cluster-operator/service/controller/internal/chartconfig"
 	configmapservice "github.com/giantswarm/cluster-operator/service/controller/internal/configmap"
@@ -40,9 +40,7 @@ import (
 	"github.com/giantswarm/cluster-operator/service/internal/cluster"
 )
 
-// ResourceSetConfig contains necessary dependencies and settings for
-// AzureClusterConfig controller ResourceSet configuration.
-type ResourceSetConfig struct {
+type resourceSetConfig struct {
 	ApprClient        *apprclient.Client
 	BaseClusterConfig *cluster.Config
 	CertSearcher      certs.Interface
@@ -62,8 +60,7 @@ type ResourceSetConfig struct {
 	ResourceNamespace     string
 }
 
-// NewResourceSet returns a configured AzureClusterConfig controller ResourceSet.
-func NewResourceSet(config ResourceSetConfig) (*controller.ResourceSet, error) {
+func newResourceSet(config resourceSetConfig) (*controller.ResourceSet, error) {
 	var err error
 
 	if config.K8sClient == nil {
