--- service/controller/azure/v22/error.go
@@ -1,12 +1,0 @@
-package v22
-
-import "github.com/giantswarm/microerror"
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
