+++ service/controller/internal/configmap/configmap.go
@@ -11,7 +11,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/key"
-	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
+	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/key"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
 	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/key"
 )
